Launch WalletActivity on Wallet Tile clicked.
If it has cards, open the WalletActivity; otherwise open the wallet app
(e.g. Google Pay).
The WalletActivity can monitor the KeyguardState- if it's locked, prompt
the user to unlock to use the wallet, and once it's unlocked, reload the
"ready to use" UI.
Also a few UI tweaks on the Wallet card carousel screen.

Test: manually test on device, see demo video
https://drive.google.com/file/d/1BzHnEnb0rWGZ1r9_go4pYgzE9A9aStDr/view?usp=sharing&resourcekey=0-Zp1277CmmrlpRGS06d395w
Test: atest
Bug: b/184490333
Bug: b/184306042
Change-Id: Ifcfb8c35ffaff3d5fcda0448f9999965c0a9159c