Don't let UdfpsView intercept touches when pauseAuth=true
The UdfpsView was still showing (with alpha 0) when
shouldPauseAuth=true, so touches weren't being sent to the window under
it. To avoid this, set the UdfpsView to invisible when
pauseAuth=true.

This CL also fixes a bug where it was possible to slide to
authenticate when auth was paused. The isWithinSensorArea method in
UdfpsController needs to have an extra check for whether auth is paused.

Test: manual, atest SystemUITests
Fixes: 187016303
Change-Id: Idcfe95e68cb5ee8e8472f1ecde2bd1fb74f9a151