Add new default vibration for notifications
Use the new WaveformBuilder to create a new default vibration waveform
that also changes the vibration frequency on devices that support it.

Create VibratorHelper in notification package to extract interactions
with the Vibrator service, the creation of VibrationEffect instances
and loading the default and fallback vibrations, using the new pattern
when possible.

Also update NotificationRecord to store a built VibrationEffect instead
of a amplitude-only waveform pattern.

Fix: 186740660
Test: BuzzBeepBlinkTest
      NotificationRecordTest
      VibratorHelperTest
Change-Id: I6c93eb729972abb7a0cfc2b6fd567606224f2f2b